refactor(frontend): ♻️ update folder to feedpage and postview
